test: split database tests into modular structure
Reorganize database_tests.rs into a modular structure:

tests/bdb.rs           - Entry point for BDB test module
tests/bdb/common.rs    - Shared test utilities and fixtures
tests/bdb/crud.rs      - CRUD tests (list, get, create, delete)
tests/bdb/actions.rs   - Action tests (export, import, backup, restore, upgrade)
tests/bdb/monitoring.rs - Monitoring tests (shards, alerts, peer stats)

Benefits:
- Clearer test organization by functionality
- Easier to locate specific tests
- Can run specific test modules (e.g., cargo test --test bdb crud::)
- Establishes pattern for splitting other large test files

This demonstrates the modular test structure proposed in issue #7.
Other test files can be split following the same pattern as needed.

Closes #7
